Code/requirements.txt
numpy==1.19.0
pillow==7.2.0
torch==1.5.1+cpu
torchvision==0.6.1+cpu